Restructure code base
