+++ pkg/util/config.go
@@ -1,4 +1,4 @@
-package utils
+package util
 
 import "os"
