+++ pkg/util/database.go
@@ -1,4 +1,4 @@
-package utils
+package util
 
 import (
 	"cloud.google.com/go/firestore"
